# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Bible Atlas API
  description: The Bible Atlas HTTP API, generated from the Rust that serves it.
  version: 0.8.0
paths:
...
  /api/chapter/{cref}:
    get:
      tags:
      - reading
      summary: |-
        `GET /api/chapter/{cref}`. `cref` must parse as exactly a
        `ScriptureRef::Chapter` (book + chapter, e.g. `EXO.14`) — a book-only or
        verse/passage-shaped path segment is the wrong shape for this endpoint
        and 400s as `bad_ref`, same as an unparseable one. The optional
        `?translation=kjv` query param (ruling 5, M1 is KJV-only) is never
        extracted, so its presence or absence cannot affect this handler at all.
      description: |-
        ruling-3-policy: once `cref` parses as a `Chapter`, an out-of-range
        chapter number (or a book with no known chapters in this atlas) is NOT an
        error — the verse-count bound comes from `canon.books[].chapters`, and an
        unknown/short chapter just yields `verse_count = 0`, i.e. a 200 response
        with an empty `verses` list. Same rationale as `scene_scripture`: a
        reader showing "no verses in this chapter" is a meaningful response, not
        a failure.

        Batch M-A (brief requirement 5, "re-implement the OLD /api/chapter
        handler as a VIEW over the window query"): the verse TEXT below now
        comes from `GraphState::chapter_span` + `GraphState::window` -- the SAME
        windowed reading-order query `GET /api/text?scope=chapter` calls --
        instead of `data.verses.get(key)`. The verse-count bound and the
        out-of-canon policy above are UNCHANGED and still sourced from
        `AtlasData`; headings moved to `graph.heading_index` in M-C2, and
        OVERLAY-1 Task 5 moved the PLACE-MENTION half onto the port too
        (`graph.scene_source(&data)`'s own `places_for_verse`/`place`, the
        materialised-from-the-graph successors of the deleted
        `AtlasData::places_for_verse`/`place_by_id` -- identical ids in
        identical order, see those methods' own doc comments). THIS endpoint
        (the reader's own chapter view) was untouched by Batch M-B's own
        event-world migration;
        only `/api/narrative/event/{id}` (see that handler's own doc comment)
        and the generic `/api/node`/`/edges` endpoints move to the graph this
        batch. The WIRE SHAPE is byte-for-byte identical -- proven by
        `tests/graph_equivalence.rs`'s own all-1,189-chapters comparison -- so
        every existing caller of this endpoint (the reader's chapter view/
        mini-reader/split view, `ChapterNode`, `PlaceCard`'s hover verse text,
        `PassageBlock`, `PopoverSectionProviders`) now serves from the graph
        with NO client-side change and no reader-visible behavior change.
      operationId: chapter
      parameters:
      - name: cref
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
  /api/contents/{corpus}:
    get:
      tags:
      - contents
      operationId: contents
      parameters:
      - name: corpus
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contents'
        '400':
          description: bad_window | bad_ref | bad_kind | bad_dir | bad_corpus
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '404':
          description: not_found
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
        '500':
          description: internal
          content:
            application/json:
              schema:
                type: object
                required:
                - error
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorInner'
                additionalProperties: false
...
components:
  schemas:
...
    Chapter:
      type: object
      required:
      - ref
      - book
      - chapter
      - verses
      properties:
        book:
          type: string
        chapter:
          type: integer
          format: int32
          minimum: 0
        ref:
          type: string
        verses:
          type: array
          items:
            $ref: '#/components/schemas/Verse'
      additionalProperties: false
    Contents:
      type: object
      required:
      - corpus
      - version
      - roots
      properties:
        corpus:
          $ref: '#/components/schemas/Corpus'
        roots:
          type: array
          items:
            $ref: '#/components/schemas/ContentsRoot'
        version:
          type: string
      additionalProperties: false
    ContentsChild:
      type: object
      required:
      - id
      - title
      - kind
      - ref
      - count
      properties:
        count:
          type: integer
          description: 'The child''s own members: verses of a chapter, paragraphs of an article.'
          minimum: 0
        id:
          type: string
        kind:
          $ref: '#/components/schemas/ContentsChildKind'
        ref:
          type: string
          description: '`GEN.1` for a chapter; `BoC 7.2.1` (its first paragraph) for an article.'
        title:
          type: string
      additionalProperties: false
    ContentsChildKind:
      type: string
      description: |-
        What a second-level entry IS: a chapter of a book, or an article of a
        document. The tree stops here (D4: "Stop at the level of ARTICLE (BoC) or
        TOPIC (Small Catechism)").
      enum:
      - chapter
      - article
    ContentsRoot:
      type: object
      required:
      - id
      - title
      - kind
      - ref
      - children
      properties:
        children:
          type: array
          items:
            $ref: '#/components/schemas/ContentsChild'
        group:
          oneOf:
          - $ref: '#/components/schemas/Testament'
            description: Absent for a Concord document, which belongs to neither testament.
          - type: 'null'
        id:
          type: string
        kind:
          $ref: '#/components/schemas/ContentsRootKind'
        ref:
          type: string
          description: 'The navigation target: the root''s own first child''s ref.'
        title:
          type: string
      additionalProperties: false
    ContentsRootKind:
      type: string
      description: |-
        What a top-level entry of a corpus's contents IS: a book of the Bible, or
        a document of the Book of Concord.
      enum:
      - book
      - document
...
    Corpus:
      type: string
      description: |-
        Every body of text this app serves a reading spine for. `id` is the
        graph's own corpus key, so the wire label and the spine key can never
        disagree.
      enum:
      - bible
      - concord
...
    Heading:
      type: object
      description: |-
        Mirrors `atlas_core::data::HeadingEntry` (event id + title + kind) plus
        ONE field it does not carry -- `is_continuation` (M-D1 requirement 1,
        CHAPTER-BOUNDARY CONTINUATION, below): the resolved pericope heading
        anchored at one verse, `is_continuation` true iff this verse is NOT the
        container's own true first-covered verse, but a LATER chapter this same
        container's coverage continues into. The atlas-core original has no
        counterpart field (this module's own doc comment already discloses why
        continuation headings are graph-side only: the atlas-core path is a
        dead, tested reference oracle for the base anchor law, never a live
        consumer this new law needs to reach).
      required:
      - event_id
      - title
      - kind
      - is_continuation
      properties:
        event_id:
          type: string
        is_continuation:
          type: boolean
        kind:
          type: string
        title:
          type: string
      additionalProperties: false
...
    PersonRef:
      type: object
      description: |-
        M-D3 (owner ruling U5, "in-text person and place name links,
        mentions-attested ONLY -- no free-text matching ever"): one person the
        graph's own `mentions` relation attests at a verse's own locus
        (`GraphService::persons_by_verse`, precomputed off `Graph.mentions`,
        NEVER a free-text/dictionary lookup against the verse's own rendered
        words -- the exact "Sin"-the-city vs "sin"-the-noun hazard this
        wording guards against: a name is only ever offered as a link because
        a CURATED edge attests that specific entity at that specific verse,
        not because some word in the text happens to match a name string
        somewhere in the graph). Deliberately lean, mirroring `PlaceRef`'s
        own "no more than the consumer needs" wire philosophy -- id (to open
        the entity's own node popover) and display label (for the client's
        own plain-text search locating WHERE in the verse's already-attested
        text this entity's own name appears, the SAME PlaceMentions.cs
        mechanism this batch generalizes rather than replaces -- attestation
        is what THIS field guards; locating the substring within an ALREADY-
        attested verse is a separate, later step, not a second matching
        hazard).
      required:
      - id
      - name
      properties:
        id:
          type: string
        name:
          type: string
      additionalProperties: false
...
    PlaceRef:
      type: object
      description: |-
        Batch R requirement 5 (place-in-verse hover -> marker blink): one place
        mentioned in a verse, per `AtlasData::places_for_verse` (the reverse of
        `Place::verse_links`). Deliberately lean -- id (to target a map marker)
        and display name (for the client's own plain-text substring match against
        the verse's rendered text, see `chapter`'s own doc comment for why there
        is no richer per-mention offset data) -- mirrors `QuietPlace`'s own
        "no more than the consumer needs" wire philosophy.
      required:
      - id
      - name
      properties:
        id:
          type: string
        name:
          type: string
      additionalProperties: false
...
    Testament:
      type: string
      description: |-
        Which half of the canon a book belongs to -- the reader's own OT/NT
        grouping, served as `ContentsRoot.group`.
      enum:
      - OT
      - NT
...
    Verse:
      type: object
      required:
      - verse
      - text
      - places
      - persons
      - xref_count
      - words_of_christ
      properties:
        heading:
          oneOf:
          - $ref: '#/components/schemas/Heading'
            description: |-
              Batch T requirement 5: this verse's own pericope heading, when it is
              a heading ANCHOR (`AtlasData::heading_for_verse`) -- omitted (not
              null), matching `Polity.transition`/`.fall`'s own conditional-
              presence convention, since the overwhelming majority of verses never
              anchor a heading at all.
          - type: 'null'
        persons:
          type: array
          items:
            $ref: '#/components/schemas/PersonRef'
          description: |-
            M-D3 (owner ruling U5): every person the graph's own `mentions`
            relation attests at this verse's own locus, in wire (row-insertion)
            order -- see `PersonRef`'s own doc comment for the full
            mentions-attested-only reasoning. Always present, possibly empty
            (most verses mention zero persons by name) -- SAME conditional-
            presence-lives-client-side convention `places` immediately above
            already establishes for this exact shape of field.
        places:
          type: array
          items:
            $ref: '#/components/schemas/PlaceRef'
          description: |-
            Batch R requirement 5: every curated place whose OWN `verse_links`
            names this verse, in place-id order (stable, not meaningful --
            there is no "primary mention" concept). Always present, possibly
            empty (most verses geocode to zero places) -- conditional presence
            lives on the CLIENT side (a mini-reader renders no hoverable span at
            all when this is empty), not as an omitted wire key, matching every
            other "always an array" field in this app's own wire (e.g.
            `Scene.quiet_places`).
        text:
          type: string
        verse:
          type: integer
          format: int32
          minimum: 0
        words_of_christ:
          type: array
          items:
            $ref: '#/components/schemas/WordsOfChristSpan'
          description: |-
            Batch RED-1: this verse's own aligned sub-verse red-letter spans
            (`GraphService.red_letter_spans`, precomputed once -- see that
            field's own doc comment for why this ONE companion is not
            graph-derivable), in ascending order. Always present, possibly
            empty -- SAME "always an array" convention `places`/`persons`
            above already establish (most verses carry none). KJV-ONLY by
            construction (decision 5: sub-verse precision is KJV-specific;
            this endpoint has never served any other translation -- ruling 5,
            "M1 is KJV-only").
        xref_count:
          type: integer
          description: |-
            Batch M-D2 (owner's cross-reference superscript directive, closed on
            the graph platform): this verse's own `cites` edge-summary COUNT at
            its TextUnit locus -- read straight off THE PORT (`GraphQuery::
            edge_summary`, design doc §5), the SAME generic query the `/api/node/
            {id}` handler itself calls (`graph::node_card`) -- never a
            second, bespoke tally. This is a genuinely NEW field (no bespoke
            predecessor existed), sourced NATIVELY from the generic contract from
            day one: the reader's own superscript markers ARE a query of the
            graph, per the design's own P1/P4 laws, just batched onto the
            EXISTING per-chapter response (one HTTP round trip for a whole
            chapter) rather than issuing 20-30 individual `GET /api/node/{id}`
            calls per chapter, which `client/IExplorableClient.cs`'s own concrete
            implementation is what a SINGLE node's lookup uses (see that file's
            own doc comment + `GraphExplorableClientTests` for the direct,
            unbatched proof the same contract is correct in isolation). Always
            present (0 for the overwhelming majority of verses, never omitted --
            the client's own superscript decision is `count == 0` -> nothing,
            never a missing-field special case).
          minimum: 0
      additionalProperties: false
...
    WordsOfChristSpan:
      type: object
      description: |-
        Batch RED-1 (owner order 2026-08-25, "Red letters on Jesus' words in
        every translation"): one sub-verse red-letter span, CHAR (not byte)
        offsets into this verse's own `text` field -- `graph.red_letter_spans`'s
        own doc comment has the full "why char offsets" (C# `string` indexing is
        UTF-16 code units; every character in this app's own KJV text is within
        the Basic Multilingual Plane, so char count == UTF-16 code unit count
        here, always). `start`/`end` are a half-open range (`text[start..end]`
        in C# `Substring(start, end-start)` terms) -- KJV display renders this
        EXACT sub-verse span; decision 5's own "ONE render rule."
      required:
      - start
      - end
      properties:
        end:
          type: integer
          minimum: 0
        start:
          type: integer
          minimum: 0
      additionalProperties: false
...